+++ src/app/components/AboutUsHome.css
@@ -8,12 +8,12 @@ body {
 }
 
 /* Container for the About Us Section */
-.about-us-container {
+.SM-about-us-container {
   display: flex;
   justify-content: center;
   align-items: center;
   /* background-image: url('/bg1.png');  */
-  background-color: #fff;
+  background-color: #fffafa;
   background-size: cover;
   background-position: center;
   background-attachment: fixed; /* Parallax background effect */
@@ -24,7 +24,7 @@ body {
 }
 
 /* About Us Content */
-.about-us-content {
+.SM-about-us-content {
   display: flex;
   flex-direction: row;
   align-items: center;
@@ -36,13 +36,13 @@ body {
   border-radius: 20px;
   box-shadow: 0 10px 30px rgba(0, 0, 0, 0.2);
   transform: translateY(-50px); /* Slight upward movement */
-  animation: fadeIn 1.2s ease-in-out; /* Smooth entrance animation */
+  animation: SM-fadeIn 1.2s ease-in-out; /* Smooth entrance animation */
   margin: 0 auto; /* Center horizontally */
   height: 500px;
 }
 
 /* Image Section */
-.about-us-image img {
+.SM-about-us-image img {
   width: 300px;
   height: 300px;
   border-radius: 15px;
@@ -50,13 +50,12 @@ body {
 }
 
 /* Text Section */
-.about-us-text {
+.SM-about-us-text {
   max-width: 800px;
   padding-left: 20px;
-
 }
 
-.about-us-text h2 {
+.SM-about-us-text h2 {
   font-size: 40px;
   color: #0A1630;
   margin-bottom: 10px;
@@ -64,7 +63,7 @@ body {
   padding-left: 15px;
 }
 
-.about-us-text h3 {
+.SM-about-us-text h3 {
   font-size: 20px;
   color: #444;
   margin-bottom: 10px;
@@ -72,7 +71,7 @@ body {
   padding-left: 15px;
 }
 
-.about-us-text p {
+.SM-about-us-text p {
   font-size: 1rem;
   color: #555;
   line-height: 1.6;
@@ -81,7 +80,7 @@ body {
 }
 
 /* Read More Button */
-/* .read-more-button {
+/* .SM-read-more-button {
   padding: 10px 20px;
   border: 2px solid #222;
   background-color: transparent;
@@ -93,14 +92,14 @@ body {
   transition: all 0.3s ease;
 } */
 
-/* .read-more-button:hover {
+/* .SM-read-more-button:hover {
   background-color: #222;
   color: #fff;
 } */
 
 @import url("https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100..900;1,100..900&display=swap");
 
-.join-us-wrapper button {
+.SM-join-us-wrapper button {
   outline: none;
   cursor: pointer;
   border: none;
@@ -117,21 +116,20 @@ body {
   overflow: hidden;
   background: #0A1630;
   color: #000000;
-
 }
 
-.join-us-wrapper button span {
+.SM-join-us-wrapper button span {
   position: relative;
   z-index: 10;
   transition: color 0.4s;
 }
 
-.join-us-wrapper button:hover span {
-  color:#fffafa;
+.SM-join-us-wrapper button:hover span {
+  color: #fffafa;
 }
 
-.join-us-wrapper button::before,
-.join-us-wrapper button::after {
+.SM-join-us-wrapper button::before,
+.SM-join-us-wrapper button::after {
   position: absolute;
   top: 0;
   left: 0;
@@ -140,7 +138,7 @@ body {
   z-index: 0;
 }
 
-.join-us-wrapper button::before {
+.SM-join-us-wrapper button::before {
   content: "";
   background: #C4C4C4;
   width: 120%;
@@ -149,13 +147,12 @@ body {
   transition: transform 0.4s cubic-bezier(0.3, 1, 0.8, 1);
 }
 
-.join-us-wrapper button:hover::before {
+.SM-join-us-wrapper button:hover::before {
   transform: translate3d(100%, 0, 0);
 }
 
-
 /* Parallax Animation (Optional Fade-In) */
-@keyframes fadeIn {
+@keyframes SM-fadeIn {
   from {
     opacity: 0;
     transform: translateY(20px);
