+++ src/app/components/HowItWorks.css
@@ -1,16 +1,18 @@
+@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap");
 
 .process-flow {
   font-family: "Roboto", sans-serif;
   text-align: center;
   padding: 3rem 2rem;
   background: transparent;
-  color: #000000;
+  color: #0A1630;
   position: relative;
 }
 
 .parallax-background {
   background: url("/path-to-your-image.jpg") no-repeat center center fixed;
   background-size: cover;
+  background-attachment: fixed; /* Ensures the parallax effect */
   min-height: 100vh;
   display: flex;
   align-items: center;
@@ -32,7 +34,6 @@
   max-width: 700px;
   margin-left: auto;
   margin-right: auto;
-
 }
 
 .flow-timeline {
