+++ src/app/components/pages/JoinUs.css
@@ -5,7 +5,7 @@
   cursor: pointer;
   border: none;
   padding: 0.9rem 2rem;
-  margin-top: 1rem;
+  margin-top: 10rem;
   font-family: "Roboto", sans-serif;
   font-size: 17px;
   position: relative;
